🐛 FIX: comments for filter examples
