+++ examples/nodejs/examples/memory/memory.retrieve.filters.And.ts
@@ -1,3 +1,17 @@
+/**
+ * Basic example to demonstrate how to retrieve memories with filters.
+ * 
+ * - And: This filter is used to retrieve memories that match all the filters.
+ * - Eq: This filter is used to retrieve memories that match the exact value.
+ * 
+ * In this example, we retrieve memories with the following filters:
+ * - company: Langbase
+ * - category: docs
+ * 
+ * We expect to get all chunks of memory from the Langbase Docs memory that have the company Langbase and the category docs.
+ * 
+ */
+
 import 'dotenv/config';
 import {Langbase} from 'langbase';
