+++ examples/nodejs/examples/memory/memory.retrieve.filters.Eq.ts
@@ -1,3 +1,15 @@
+/**
+ * Basic example to demonstrate how to retrieve memories with filters.
+ * 
+ * - Eq: This filter is used to retrieve memories that match the exact value.
+ * 
+ * In this example, we retrieve memories with the following filters:
+ * - company: Langbase
+ * 
+ * We expect to get all chunks of memory from the Langbase Docs memory that have the company Langbase.
+ * 
+ */
+
 import 'dotenv/config';
 import {Langbase} from 'langbase';
