+++ examples/nodejs/examples/memory/memory.retrieve.filters.In.ts
@@ -1,3 +1,15 @@
+/**
+ * Basic example to demonstrate how to retrieve memories with filters.
+ * 
+ * - In: This filter is used to retrieve memories that match any of the value/values in the array.
+ * 
+ * In this example, we retrieve memories with the following filters:
+ * - company: Langbase or Google
+ * 
+ * We expect to get all chunks of memory from the Langbase Docs memory that have the company Langbase or Google.
+ * 
+ */
+
 import 'dotenv/config';
 import {Langbase} from 'langbase';
